Add wallet data to context & split up handlers
